requirements.txt
fitz
openai
requests
azure-storage-blob